fixes #10033 PROBLEM: The Linkgrabber Sidebar contains filters. If we click on or several filters, affected links in the table get selected.
If a package contains a link that should get selected, the package got selected instead.

All Actions work on the selection. if a collapsed package is selected, all it's children are part of the selection.

This is a conflict with the sidebar filters.

SOLUTION:
To get a solution, we now expand all packages that contain links to select.